feat: add option for custom hash lib
This enables providing a custom hash library for hashing
and verifying tokens. One needs to override a function called
'hash' and another function called 'verify'.
If no such object with these functions is provided, it defaults
to the bcrypt hashing library.

In the tests of this library the hashing lib Argon2 is used
(see https://github.com/ranisalt/node-argon2)